Adaptations for macOS 10.13 and Xcode 9.
In particular,  macOS 10.13 and iOS 11 change the SceneKit nodes hierarchy for reference nodes: the root node of a referenced scene is no longer a child of the corresponding SCNReferenceNode, but is now considered to be the SCNReferenceNode itself.
To stay compatible with previous OS versions, the reference node for collected objects is now determined dynamically by searching a SCNReferenceNode ancestor of the collided SCNNode.